requirements.txt
asgiref==3.8.1
Django==5.2
django-tinymce==4.1.0
pillow==11.2.1
psycopg2==2.9.10
setuptools==79.0.1
sqlparse==0.5.3
wheel==0.45.1
